feat: benchmark tx size with fee (#5414)
This PR adds new benchmarks to compare tx size with different fee
payment methods.

The tx type chosen for this benchmark is a simple private transfer:
`token_contract.transfer(alice, bob, 1n, 0)` . This is expected to (1)
nullify one of Alice's note (2) create a note for Bob with the amount
and (3) create a note for Alice with the left over from the nullified
note. On top of this base we add the costs of paying the fee in
public/private

Fix #5403