chore: add goblin ops in add_gates_to_ensure_all_polys_are_non_zero  (#5468)
Resolves https://github.com/AztecProtocol/barretenberg/issues/843.

Removes the need to add golin ecc ops  for each UGH circuit by
ensuring we add the dummy gates as part of the function used to ensure
other polynomials are non-zero due to the absence of specific gates.
However, we keep adding goblin ecc gates
to circuits in situations when we want to test Goblin and also in
ClientIVC because merge proving is done prior to ProverInstance
creation.