feat: Brillig pointer codegen and execution (#5737)
Resolves https://github.com/noir-lang/noir/issues/3907

This PR builds upon
https://github.com/AztecProtocol/aztec-packages/pull/5709.

These changes do not yet include a Brillig stdlib and removal of the
`Brillig` opcode itself. The generated stdlib Brillig (such as for
quotient) is not created in the same manner as other Brillig calls which
are generated during SSA. I have decided to leave this for another
follow-up where we can actually remove `Brillig`.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: jfecher <jfecher11@gmail.com>