feat: add the storage layout to the contract artifact (#5952)
Fixes #5947 and includes the notes as well. The notes was part of the
contract object, but not directly exposed on the artifact itself.

Needed to fix a few additional things to make it work:
- The `deploy_contract` function was broken (old call flow), it cannot
do the new flow because that would be a circular dependency.
- Using the `storageLayout` values to get the `storageSlot` that we are
using in multiple tests
- Using the `notes` values to get the `noteTypeId` that we are using in
multiple tests
- Removed `lodash.uniqby` as the dependency was not needed after this.