feat: Private Kernel Recursion (#6278)
This PR introduces recursive verification to the private kernel
circuits. Both app circuit and previous kernel circuit proofs are
verified. This closes #5978

The changes can be largely categorised as:

1. PXE modifications to pass proofs and verification keys from the
output of a proving process as inputs to the next simulation/proving
process.
2. Serialisation of `PrivateCircuitPublicInputs` and
`PrivateKernelCircuitPublicInputs` structs to fields.
3. Aggregation of proofs using Noir's `verify_proof` api.

Additional task create
[here](https://github.com/AztecProtocol/aztec-packages/issues/6285) to
prevent the specification of `pub` on arguments to private functions.