feat: Update the encrypted note log format (#6411)
Fixes #5901.

Uses the new format to match the keys and logs spec. The current
implementation here is still using a typescript implementation to
encrypt the data, mainly using the new flow.

The intention is that #6408 and #1139 can be addresses separately and
than we can just replace the import to use the constrained version
instead of the oracle at that point.

Note, that the outgoing logs are currently less than meaningful, as some
of the infrastructure is not yet in place to handle those nicely, see
#6410 for more on that.

What was called the encrypted_log_payload in #6348 have been moved into
the l1_payload, to better integrate with the rest of the setup.