chore: sort public call requests in the circuits (#6650)
This removes some patching we were doing in TS around the public call
stack items.

Now, we follow the existing pattern to sort public call stack items in
TS, then assert the sort in the circuit.

The only difference is that call *stack* items need a reverse sort, so
the one with the lowest side effect counter is last (and so is run
first).

Also, started making circuit types in TS readonly.