feat: Add code-workspace and update build dirs (#6723)
Adds a set of VS Code settings in proving-systems.code-workspace akin to
what we have in barretenberg.code-workspace, but situated at the top of
the monorepo so that it's easier to work across the full stack. It also
changes the build dir for debug and with-asserts builds to give a better
incremental workflow.