chore: renaming private token as token with refunds (#7626)
Private token makes me think that the token is working only with private state so the name is confusing. Also I plan on using the reference token impl and just extending it with the refund stuff in a future PR. This makes sense because a lot of the custom changes to the token were made because npk was used instead of owner address --> this required modiying balances_map. Given that this will be tackled the standard way now that we are able to compute slotted hiding point in private I can revert most of the custom changes to the token.