feat: some fixes and cleanup in PG recursive verifier (#8053)
Some minor reorganization/fixes for VerifierInstance/VerificationKey
constructors. There are lots of places where we are manually reproducing
something resembling the VerificationKey constructor. This PR tries to
simplify some of that logic where possible and adds in handling of
components of the vkey that were missed in a few of these locations.
(Done in prep for introducing PG verifier logic based on stdlib types).