fix: use different rust toolchain for foundry builds (#8869)
## Overview

Cargo version installed by default in `build-images` was not the minimum
required for the newer release of foundry that was pinned in
https://github.com/AztecProtocol/aztec-packages/pull/8868, this alters
the `foundry-build` earthly script to use `1.80` rustc version. This is
set through adding a toolchain file so the correct binary will be
installed automatically.

In the future if we need to bump the version for builds, we can just
alter this toolchain version.

This also runs forge fmt for this updated version