refactor: modify note processors and synchronizers to use complete address (#9152)
I am modifying the note processors and synchronizers to use a complete
address, instead of purely using the aztecaddress. This is because in
the new address scheme, to decrypt, we will require the preaddress (a
future member of the complete address), as well as the ivsk to calculate
the "secret" of the address point. This is just scaffolding to be able
to make the decrypting changes in the future.

The note processors specifically need this, but the synchronizer
previously did not have it, but it was fetching the complete address
anyways from the pxe db.