feat: reorder blocks for efficiency (#9560)
Reorders blocks (Mega only) so that data bus reads are close to the top
of the trace and lookups are at the bottom. This helps minimize both the
memory required to store the log-deriv inverse polynomials as well as
the "active" region of the trace by minimizing the distance (maximizing
the overlap) of the various lookup gates with the data from which they
read. (Note: tables are constructed at the bottom of the trace).