feat: browser chunking (#11102)
Further modularises `noir-protocol-circuits-types`, exposing a
`client/lazy` API that allow browser bundles to not import every single
JSON artifact at once, but rather pull them on demand. `/server` and
`/vks` entrypoints are also provided to aid in startup CLI times, jest
cache and general cleanup.

This is very important due to the sheer amount of different reset
variants we currently have. However, a regular `/client/bundle` is still
exposed in case dynamic imports are not supported in the target
platform.

This takes our bundle sizes in the example vite box (with PXE in the
browser and proving) from 50MB (30 compressed) to ~3MB before meaningful
content is displayed. After that, `bb.js` is an additional 8-9 and each
circuit hovers around 0.3-1, but only the ones used in the transactions
are pulled when required.

Also removed the `TestPrivateKernelProver`, and replaced it by each of
our prover implementations providing a cleaner simulation vs.
witgen+proving interface. This way, the prover is now 100% a PXE plugin
(which will allow us to even delegate proving!)