feat: Validator deadline for reexecution (#11050)
Adds a reexecution deadline for validators, so they abort trying to
reexecute public txs while block building if too late into the slot.

As part of this PR, we are adding a hook to tweak logger module names
(so we can inject an id when running multiple components in the same
process), and extracting the chain monitor from the e2e epochs test.

Fixes #10959