fix: Fail in proxy deployment should fail the step (#11308)
Running `cast` in the conditional caused bash to swallow any error
connecting to the ETHEREUM_HOST, so the step was not retried, so we
proceeded without a create2-deployer, so the deploy-l1-contract step
next would never succeed.

Moving it outside the conditional ensures that the exit code is
propagated, so kind should retry the step (I hope!).