chore: Trace propagation from json rpc client to server (#11325)
Uses otel's `propagate` API to push the current trace from the json rpc
client to server using http headers. Client is instrumented by wrapping
the `fetch` function so it creates a new span and creates the headers,
and the server is instrumented by injecting new middleware that captures
the headers and starts a new span for the scope of the request.

Fixes #11185
Fixes #11223

Cherry-picks work by @alexghr from #11222

### Example run

Note that `traceId` is the same in both processes.

Bot with pxe logs:
```
[20:08:54.763] INFO: pxe:service Sent transaction 0x125eb560f33f44a28b98f0f065197eed8a8557bcb1ab2659eb34e5be0b33a9b4 {"trace_id":"d1143080037e46c037dea7cbdf8e29d6","span_id":"236692ffb270ecb4","trace_flags":"01"}
[20:08:54.763] INFO: bot Sent tx with hash 0x125eb560f33f44a28b98f0f065197eed8a8557bcb1ab2659eb34e5be0b33a9b4 {"trace_id":"d1143080037e46c037dea7cbdf8e29d6","span_id":"236692ffb270ecb4","trace_flags":"01"}
```

Node logs:
```
[20:08:54.762] INFO: node Received tx 0x125eb560f33f44a28b98f0f065197eed8a8557bcb1ab2659eb34e5be0b33a9b4 {"txHash":"0x125eb560f33f44a28b98f0f065197eed8a8557bcb1ab2659eb34e5be0b33a9b4","trace_id":"d1143080037e46c037dea7cbdf8e29d6","span_id":"5d1b907122754832","trace_flags":"01"}
```

---------

Co-authored-by: Alex Gherghisan <alexg@aztecprotocol.com>