feat(avm): interactive debugger (#11477)
Debugger plus some fixes. This has bb-pilcom changes so update and rebuild after it merges.

<div>
    <a href="https://www.loom.com/share/1ce400f55d4a4d888000cb54c7361a6f">
      <p>AVM interactive debugger - Watch Video</p>
    </a>
    <a href="https://www.loom.com/share/1ce400f55d4a4d888000cb54c7361a6f">
      <img style="max-width:300px;" src="https://cdn.loom.com/sessions/thumbnails/1ce400f55d4a4d888000cb54c7361a6f-68caa0ac8f8e7ebb-full-play.gif">
    </a>
  </div>

Note: it does not support history or arrow keys or tab/autocompletion. Mostly because this is terminal-dependent and I don't want to pull in a console dependency. I might attempt to do it from scratch when I have some free time.