chore(avm): disconnect VM1 (#12346)
![unnamed-1.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/mO08Q17TCoHzVKYudUJh/bb553009-1c23-4bc3-a5de-bcf2b886652a.png)

We are currently working on VM2 and having to maintain backwards compatibility with VM1 has reached a tipping point. This PR disconnects VM1 from the native BB binary and makes it always report success.

As a result
* `bb-prover` tests don't mean a thing (except the v2 one)
* the proving part of e2e_full_proving becomes meaningless
* **We will not be proving in any dev/testnet until we connect vm2!**
* This doesn't change recursive verification in the base rollup because we were already not doing that (too expensive until Goblinized).

However, this should let us move faster and we should be able to connect vm2 "soon".

Having VM1 connected is currently blocking work on changing the hinting schema etc.