feat: Get blobs from blob archive if not found in blob sink (#12498)
Blob sink now falls back to a blob archive source to fetch a requested
blob if not in local storage. Blobs are deserialized before storing, so
we validate that they correspond to an Aztec tx (though they may not be
from the same L2 chain being served by the blob sink, but I think that's
acceptable).

For now we rely on blobscan API for fetching these archives. The
fallback can be set manually via a `BLOB_SINK_ARCHIVE_API_URL` env var,
and is set automatically if L1 chain id is mainnet or sepolia.

Note that blobscan sepolia has stopped syncing since the Pectra upgrade
yesterday, but hopefully they'll be back up soon.