fix: Specify block number when fetching instances (alternative) (#12709)
We need to specify a block number when fetching contract instances since
during block building, we want to fetch what the instance will be for
the blocknumber being built, not the one currently synced.
Alternative to
https://github.com/AztecProtocol/aztec-packages/pull/12688 I think this
option is worse conceptually altough it's less changes. `class
WorldStateDB extends ContractsDataSourcePublicDB` is a block level cache
so it makes sense that it receives the block it's caching for on
construction. This PR requires passing block number on every fetch