fix: no hardcoded versions in bbup (#12944)
I've updated bbup to pull the version of bb to install for a given noir
version from a json file in aztec-packages (previously it would read
some file inside of the noir repo at the given release tag of noir that
they have installed.

We can then update this in future without needing to migrate people onto
new versions of bbup.