fix: use version from registry for rollup instead of config (#12938)
This ensures that we use rollup / registry from the version passed in,
and the version is now optional and not preset to 1. I'm not removing
any config though to keep this individual change small, as the cleanup
will be bigger and may want to be thought out a bit more.

Co-authored-by: sklppy88 <esau@aztecprotocol.com>