feat(avm): standard AVM recursive verifier v2 (#13234)
Essentially a port of vm1. Public inputs are validated but we support
only an extremely simplified version (just reverted flag).
Standard version means that it is not the goblinized one.
SRS needed to be increased to 2^24 as AVM2 has more columns than V1 and
the recursive verifier has about 9 millions ultrahonk gates.