chore: add default native proving for cli-wallet (#13129)
Okay so I've verified [here](http://ci.aztec-labs.com/38a5491ed3a6ac97),
that none of the cli-wallet tests in

cli-wallet/tests/flows or
aztec-up/test or
end-to-end/src/guides

are running with prover. It turns out we needed to export the var in
each location as the command was being called and not inheriting a
non-exported env var.

Co-authored-by: sklppy88 <esau@aztecprotocol.com>