feat: Honk will always have pairing point object (#13217)
Closes https://github.com/AztecProtocol/barretenberg/issues/1315.

Previously, a pairing point object may not exist in the public inputs of
a honk proof IF the recursive flag is not on. We should just simplify
things and always attach it for UH. I originally didn't always attach it
because it's slightly cheaper to not include it, but at this point, its
not worth the complexity.

It should be true that the recursive flag is NEVER used for Honk.

There may still exist plenty of code that can be cleaned up as a result
of this PR.