chore: Fix flake in reqresp p2p test (#13333)
Issue seems to be that the txs get sent to the first proposer with
little time for it to build the block, so it misses its shot, and then
that tx never reaches other nodes (since p2p gossip was disabled). This
attempts to fix it.

WIP: currently running into a `Assertion failed: Failed to get a note
'self.is_some()'`