chore(docs): bugfix + deploy on master (#13717)
# Automatically release docs on master branch

This PR adds automatic documentation release when changes are merged to
the master branch. It:

- Updates the main bootstrap.sh to call release-docs when on the master
branch
- Renames the docs release command from "release" to "release-docs" to
avoid it running on releases (we don't need it)
- Passes CI parameter to project bootstrap scripts in the release
workflow to avoid the yarn lock problem
- Removes stderr redirection in docs deployment so we can see the output
of the release in the logs
- Removes the deploy-doc github-action since that will be dealt with on
the `ci`

The changes ensure our documentation is automatically deployed whenever
changes are merged to master, improving our documentation workflow.

---------

Co-authored-by: Maddiaa <47148561+Maddiaa0@users.noreply.github.com>