feat!: improving perf insights + avoid simulating on proving (#13928)
Introduces new timing info in`profileTx`, but in the process got
distracted by a few things, so I'm branching this off before adding more
benchmarks:

- Adds a contract deployment benchmark to our key flows
- Avoids resimulating kernels (in brillig) before proving. We're going
to do witgen anyways! This should improve our performance across the
board. Simulation is still recommended, but that's left as a wallet
responsibility
- Allows playground to profile txs taking into account fee payments
and/or authwits

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>