feat: remove extra pairing point aggregation batch_mul (#14219)
Closes https://github.com/AztecProtocol/barretenberg/issues/1325.

We are currently doing pairing point aggregation with a batchmul of size
1 in operator* and then another batchmul of size 2 in operator+. If we
just directly do it, we can do it with just 1 batchmul of size 2. This
drops the number of transcript rows in the ECCVM, but doesn't affect the
actual number of rows because the MSM builder row count dominates it by
a significant margin.