fix!: Handle L1 reorgs on cross-chain messages (#14151)
## L1 contracts

- Adds a `rollingHash` computed over all L1 to L2 messages as they are
received, which gets emitted with each message and kept in storage.
- Adds a `getState` method that returns both rolling hash and total
count of messages, used for syncing.

## Message store

- Changes keys from strings to numbers so they are properly ordered
- Checks message consistency on insertion (indices, rolling hashes, l2
block numbers, etc)
- Adds method to iterate over arbitrary ranges of messages
- Stores message metadata (l1 block number, l2 block number, rolling
hash, etc) along with each leaf
- Deletes unused in-memory message store

## Archiver

- Compares local messages rolling hash vs the one on the Inbox contract
to determine if it needs to resync.
- Before syncing, checks if the rolling hash for the latest message
downloaded matches; if not, it considers it a reorg and rolls back
messages until a common sync point.
- Once syncing is completed, checks the resulting rolling hash with the
one obtained, and warns on mismatch.