feat!: minimum committee size (#15019)
Now the rollup insists that it has at least the targetCommitteeSize in
validators, else it is impossible to propose blocks because a committee
will not form because it will not sample the validator set.

The important caveat is that if the rollup is deployed with a target
committee size of 0, in which case the committee is guaranteed to be
empty, and the `ValidatorSelectionLib` `verify` function will always
pass.

Indeed, operations which query L1 for the current committee, such as
`getCommitteeAt` and `getCurrentEpochCommittee` will revert with
`ValidatorSelection__InsufficientCommitteeSize`.

Because of this, typescript needed to change. In `rollup.ts`, we
explicitly catch that error and return `undefined`. In the EpochCache,
when we explicitly ask for the next proposer, we return `undefined` if
the committee is empty, and throw a `NoCommitteeError` if the committee
does not exist on L1.

TS tests now generally specify target committee size of 0. The exception
is the e2e tests. It is now important that no transactions are awaited
before the validator nodes are up, because the initial node cannot
submit blocks.