feat: Validate block proposal in hash (#15146)
When validators attested to a block proposal, we assumed that they had
the correct set of L1 to L2 messages, downloaded from L1. But if their
archivers had not properly synced them, then the attestation would fail
with a reexecution error, caused by a mismatch in the archiver root.

This PR adds an early check for the content commitment `inHash` (should
we review this name?) in the proposal, which is a merkle root of the
L1-to-L2 messages present in the block. If the `inHash` in the proposal
does not match a value manually computed from the validator's L1-to-L2
messages for that block, we fail early rather than reexecute everything.

This change required removing the `L1ToL2MessageSource` out of the block
builder. Now, the block builder receives the l1 to l2 messages, same as
it received the txs, from its caller. Also, this PR adds extra info to
the `ReExStateMismatchError`, so we can compare individual tree roots to
spot the issue in reexecution.