feat: run nightly tests in GKE (#14417)
This PR moves the nightly tests from being run locally, in Kind, to
running against a remote GKE cluster.

Right now each test deploys a whole network. This is slow and could
cause contention if a bunch of pods fight over the same physical nodes
(and autoscaling takes time).

Next item on the list: share a single deployment between all tests.

**Docker release change:** nightly runs will push the image early
(tagged with `$GIT_SHA`) to docker hub in order for GKE to have access
to it. After all tests pass the image is tagged with semver.

Fix #13514

---------

Co-authored-by: PhilWindle <philip.windle@gmail.com>