feat: Archiver does not sync blocks with invalid attestations (#15896)
Archiver now checks committee attestations and refuses to sync a block
if it does not pass validation.

Note that this addresses scenarios where the proposer is malicious, but
does not handle cases where the entire committee is and produces
signatures for a block with an unattested parent. That'll be left for a
future PR.

Builds on #15813